// src/app/core/api/generated/model/allowanceCalcMethod.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum AllowanceCalcMethod {
    NUMBER_0 = 0,
    NUMBER_1 = 1,
    NUMBER_2 = 2
}