// src/app/core/api/generated/model/allowanceTypeDto.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AllowanceTypeDto { 
    id?: string;
    name?: string | null;
    calcMethod?: number;
    value?: number;
    isActive?: boolean;
}